breakdown template instantiator into smaller submodules
